DeskTop: Make drag selection more responsive
Replace more "map icons one by one" logic (which has significant
overhead) with remapping all icons in a window. And do CheckEvents
when rejecting an icon to keep the cursor responsive, e.g. when
there are a hundred files in the window.

Factor out a helper for mapping an individual point, to save space.