MGTK: Make SetCursor to same address a no-op
This removes the flicker associated with calling SetCursor if the
cursor address is the same, including for system cursors. This
eliminates the need for clients that potentially change the cursor on
mouse move (e.g. when there is a line edit input field and the cursor
should switch from pointer to I-beam or vice versa) to track the state
to avoid the call if there is no change.