MGTK: Reduce cursor flicker by waiting for VBL
When moving (or changing) the mouse cursor, wait for the vertical
blanking interval before doing the restore-background-then-paint
work. This eliminates flicker when just moving the mouse around.

It doesn't prevent the flicker associated with hiding/showing the
mouse around drawing operations, e.g. painting icons in a window.
Those appear to be too slow to fit into the blanking interval.