MGTK: Eliminate cursor flicker when moving mouse
When moving the mouse cursor, the previous needs to be restored
then the mouse drawn in the new position. This leads to a "flickering"
effect as the mouse is moved.

Address this by waiting for the vertical blanking interval (usually
abbreviated VBL or VBI) before doing the restore-background-then-draw
work. This eliminates flicker when just moving the mouse around.

It doesn't prevent the flicker associated with hiding/showing the
mouse around drawing operations, e.g. painting icons in a window,
which are too slow to fit into the blanking interval.

VBL is handled on the IIe, IIc/IIc+ and IIgs and compatibles with a
new WaitVBL proc. The rest of the change is reworking and optimizing
the cursor erase/redraw code to fit within the 4550 cycle blanking
interval on NTSC. The code was split into a "pre-draw" phase which
calculates what to draw and where, including the shifted cursor bits
and mask. This is run before waiting for VBL, and then erase/draw can
happen. The changes are a balance between readability, size and
performance - while keeping the erase/draw under the cycle budget, the
usual "save bytes for future features" guidance holds.

Co-authored with Colin Leroy-Mira <colin@colino.net>