MGTK: Prevent memory corruption if cursor extends above screen
Loop needs to terminate if the hotspot is in the middle of the cursor
and the cursor is near the top of the screen. With the previous code,
a bogus video pointer was computed and "interesting" things happened
to memory. Due to the way save/restore works this may not have been
detectable in most cases, but that's just luck.