DeskTop: Fix display of window headers when partially offscreen
The header areas don't scroll with the window contents and the
positioning of the "items", "in disk" and "available" labels is
dependent on the window width, so some math needs to be done to
account for the window size and scroll position. Unfortunately, this
was done using the window viewport, rather than the window properties.
This meant that when the window was partially offscreen to the left or
right the cliprect width was used rather than the window's actual
width and the cliprect position rather than the window's maprect
(viewport) position.

Make the calculations correctly use the window's maprect so that
everything displays as expected in these cases.

Judicious use of the zero page means we also shave 8 bytes.