MGTK: Add DrawString and StringWidth calls
Inspired by the IIgs ToolBox, add MGTK calls that support using Pascal
(i.e. length-prefixed) strings directly rather than requiring a
structure that gives the length separately, as in DrawText and
TextWidth, although those are retained.

This reduces the need for helpers in most cases; the DrawString call
takes the string address directly, rather than a parameter block. If
the string address is only known at runtime (i.e. dynamically
selected) then self-modifying code is still simpler than a helper in
most cases.

For the StringWidth call a parameter block is still needed, so helpers
that take the string address in A,X and return the width in A,X are
common, but no longer omnipresent.

Most call sites are updated to use the new calls rather than DrawText
and TextWidth, although those are retained and still use anywhere
there is tricksy code.

This generally saves code; an MGTK call is 6 bytes and an A,X call is
7 bytes, so that's one byte per. Cases that require SMC can be larger
although when helpers can be deleted this still saves memory. The big
win is readability.

No functional changes.