DeskTop: Do window close animation after clearing updates
Previously, the window close animation was run as soon as the window
was closed (and the desktop pattern repainted) but before any
overlapping windows were drawn. This meant that the desktop pattern
stayed visible while the animation ran, which was ugly.

Flip the order so that any necessary window repaints ("updates") are
run first. This delays the animation but that's a good thing - if
there are many intermediate windows they can take a while to repaint,
and showing the animation at the end reminds the user of what they
just did.

This was already when the animation Desk Accessories was performed,
so this is more consistent overall too.