DeskTop: Horizontally center file icons
When populating directory windows, DeskTop already ensured that icons
on the same row were vertically aligned at the bottom regardless of
the height of the bitmap. But icons in the same column were
horizontally aligned on the left edge. Some icons (notably the "app"
icons) are wider, so this was ugly.

Fix it!

Slightly non-trivial since a file's icon can change when renamed, so
we need to do the magic dance there too.