BTK: Improve visibility of BTK::Flash using VBL
Loop for 5 VBL cycles (so ~1/12 second).

Also, make Calendar DA use BTK::Flash rather than bespoke inversion.
Which turned up a minor bug in BTK that the inversion depends on the
window's default pattern, so ensure a solid pattern is always used so
that the flash is guaranteed to be visible.