Alert: Preserve cursor across calls
Previously, a call to Alert would exit with the cursor reset to the
normal pointer cursor. If called when the cursor was something else -
e.g. a watch cursor, during a long operation - it was the caller's
responsibility to restore it, e.g. if the long operation is resumed.

Flip that around; the Alert call will now restore the cursor to
whatever it was when called. Now the caller needs to manually reset
the cursor to the pointer if not desired, e.g. if the long operation
was canceled.

This makes the behavior much more obvious, allowing auditing of cursor
changes, and encourages code patterns that ensure all return paths
reset the cursor appropriately.

Shaves bytes here, costs bytes there. But it's better.

A couple of potential edge cases where the cursor wasn't reset are
fixed, and more places that probably should have shown the watch
cursor for consistency are tweaked.