Add MAME+Lua based testing infrastructure
Take advantage of MAME's support for Lua scripts to drive the emulator
to implement automated testing.

* Tests live in tests/ and are Lua scripts Tests are run with
* bin/mametest PATH_TO_SCRIPT.lua

Expect plenty of iteration on the tooling, scripts, etc. The goal is
to automated as much as possible from notes/testplan.md so that it
takes much less time to validate a release candidate.

Right now, the test infrastructure mostly just drives A2D and dumps
out snapshots at key points, requiring the user to verify that
everything looks correct. This is a long way from CI/CT but is still
of tremendous benefit.