DeskTop: Don't show error checking all drives with empty floppy
When polling Disk II devices, ERR_DEVICE_NOT_CONNECTED is returned by
the ProDOS driver if there is no disk in the drive. This case wasn't
explicitly handled by Special > Check All Drives (or Special > Check
Drive) so an error was shown. Filter that out.

Doesn't repro in Virtual ][ but does repro in MAME and this looks like
a real issue, not emulator.

Probably regressed with the rework after the 1.5 branch.