MGTK: Don't invoke first menu item without shortcuts on ^@
Control+Shift+2 a.k.a. ^@ yields the ASCII NUL value. This was used in
the menu definitions to signal "no shortcut". MGTK didn't special-case
this, so the first menu item without a shortcut would be invoked when
this key combination was pressed.

In DeskTop this was "About Apple II DeskTop" so the dialog would show!

Address this by having the menu defining macros put a value with the
high bit set; MGTK strips the high bit from the keyboard input so it
can never match. This allows MGTK apps to actually define a shortcut
with ^@ if they want to.