MGTK: Add automation interface
Add a mechanism for test automation to interact directly with MGTK,
allowing queries like "how many windows are open?", "what is the name
of the top-most window?", and "where is the top-most window's title
bar?" Since the MGTK API is relatively stable this is a safe
integration point.

Several tests updated to take advantage of this, but more can be
done.