Tests: Package results into a ZIP file
Since each run empties the snaps/ dir, running mametest in a loop will
not leave useful images behind except for the last. This makes
mametest drop a file in mametest/results named for the test, e.g.
alerts.lua produces alerts.zip. The stdout/stderr output of the run
is also captured as an "output.txt" file in the ZIP.