MGTK: Change MouseKeys to make dragging easier, support modifiers
Previously, in MouseKeys mode the Solid Apple key was used as the
mouse button. This was easy to implement as it could be held down
while other keys were used to perform drag operations. Unfortunately,
this conflicted with the use of the key as a modifier for operations,
so you could not perform a drag/drop while holding Solid Apple to
force a copy (same volume) or move (other volume). This impacted
automated testing, which uses MouseKeys extensively.

This is now changed to be similar to how modern macOS works, albeit
with different keys:

* OA + SA + Space enters MouseKeys mode (no change)
* Arrow keys move the mouse cursor (no change0
* Space now clicks the mouse button
* Comma now presses and holds the mouse button (start a drag)
* Period now releases the mouse button (end a drag)
* Escape exits MouseKeys mode (no change)