DeskTop: Animate from just icon bitmap, not total bounds
When window open/close animation was updated in 90a46a5 to use a
starting rect not just a point at the icon, there wasn't a convenient
way to get just the icon bitmap bounds, so the whole icon bounding
rect (including the label) was used. There's an API for that now, so
switch over, as that matches the classic Mac OS appearance.