DeskTop: Fix error shown when directory can't be deleted
ProDOS reports an access error so we'd show "this is locked, want to
unlock?" alert. But this is not helpful if the error was from another
reason, e.g. the directory wasn't empty (e.g. contained a GS/OS forked
file we couldn't remove). So show a generic error instead.

Also, tests.