DeskTop & Tests: Coverage for file deletion, fix recursive deletion
Previous fix in 28bfb50 was correct - don't prompt for unlock if not
locked - but the file info snapshot was not correct after finishing
directory enumeration and trying to delete the directory, so it gave
false-negatives. This was probably due to recent work to minimize the
number of GET_FILE_INFO calls. Anyway, they're needed here so slip
them in where needed. Yay for tests.