DeskTop: Fix hang after non-menu shortcut, if window fails to open
When a window fails to open (e.g. the disk has been ejected) then the
work unwinds via a saved stack pointer. That was set up correctly for
actions dispatched via clicks and via menu and menu-based shortcuts.
But for actions with dual modifiers like Open Apple + Solid Apple +
Down the stack pointer was not saved, so recovery could go off into
the weeds based on previous actions.

Move the stack pointer saving up to the start of keyboard handling.

I couldn't repro this manually but fortunately it turned up in
a MAME-based test script.