DeskTop: Fix clipping of desktop icons on Select All
When 8a35bcd added support for Select All when the desktop had focus,
this broke an assumption that only the top-most window (or desktop if
none) would need a repaint. This had allowed the IconTK::DrawAll fast
path to be used, which never clips. But with the change, this meant
desktop icons wouldn't be clipped against overlapping windows. Oops!

Switch to using the existing RedrawSelectedIcons call which already
has the smarts to use a fast-path when possible, and also keeps the
cursor responsive. So it saves bytes - yay!

This potentially regresses performance slightly for the redraw for
Select All in a window because redraw iteration now takes place within
DeskTop rather than IconTK, but the logic is very similar in both
places (clipping, etc) so it is likely minor and we can revisit in the
future if needed.