Sci Calc: Fix incorrect ASIN results
Change 0c8c09a shifted the ASIN calculation function up by 5 bytes,
which placed the PushARG call to a page boundary, leading to it
exiting with A=0/Z=1. This caused the subsequent FMULTT to do the
wrong thing - it was expecting Z set based on the previous FAC update.
D'oh!

Fortunately this was caught in a test. The fix is to expand a function
called before FMULTT (and FDIVT) to correct states, exactly like this.
Instead of preserving A it now sets A (and Z) as expected for those
calls.