DeskTop: Fix behavior after File > Duplicate failures
When File > Duplicate failed due to lack of space or unsupported
storage type it would (1) show an OK/Cancel alert (which is silly as
it can't continue) and (2) if OK was welected it would prompt to
rename the previous file. Not good!

Fix this so that when one of these errors is showing during copy, if
it's the last file then an OK-only alert is shown, and it is treated
as-if Cancel was selected. This affects more than Duplicate but it
should be in a good way.