Tests: Add File Picker test for expected blank Dir/Disk names
While writing tests I made a note that the behavior of the file picker
differs between starting with an empty path (e.g. just volume
selected) vs. navigating to an empty path, where in one case the boot
vol is shown for both dir and disk names where the expectation is that
they are blank. But I can't reproduce it. Adding the test for
posterity.