DeskTop: Init maprect for empty window to avoid scrollbars on restore
When an empty window was created, the maprect was set with origin 0,0
which is fine as there are no icons. If a file is added within the
DeskTop then the maprect is refreshed based on the icon bounding box,
and scrollbars are avoided if possible.

But if DeskTop quits and a file is added, when the window is restored
the bounding box is not adjusted so the icon is clipped, with
scrollbars appearing. This is probably new in v1.5 as in prior
versions the icons were placed with a horizontal offset, whereas in
v1.5 this was dropped and we just relied on the bounding box.

Avoid this by setting a good default offset. Only really works for the
single icon case, but more icons than that will be clipped anyway if
the window size is small.